requirements.txt
asyncpg
discord.py
timeago
jishaku
python-socketio
typing_extensions